# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Investintell Light API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/funds": {
      "get": {
        "tags": [
          "funds"
        ],
        "summary": "List Funds",
        "description": "One page of the fund universe with headline risk metrics.\n\nRisk-metric bounds drop funds lacking that metric by definition.",
        "operationId": "list_funds_funds_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 80
                },
                {
                  "type": "null"
                }
              ],
              "description": "Ticker/name substring match.",
              "title": "Search"
            },
            "description": "Ticker/name substring match."
          },
          {
            "name": "fund_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "etf",
                    "mmf",
                    "mutual_fund"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fund Type"
            }
          },
          {
            "name": "strategy_label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 80
                },
                {
                  "type": "null"
                }
              ],
              "description": "Strategy label substring match (free text).",
              "title": "Strategy Label"
            },
            "description": "Strategy label substring match (free text)."
          },
          {
            "name": "asset_class",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "equity",
                    "fixed_income",
                    "cash",
                    "alternatives"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Class"
            }
          },
          {
            "name": "expense_ratio_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expense Ratio Max"
            }
          },
          {
            "name": "aum_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aum Min"
            }
          },
          {
            "name": "sharpe_1y_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sharpe 1Y Min"
            }
          },
          {
            "name": "volatility_1y_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Volatility 1Y Max"
            }
          },
          {
            "name": "return_1y_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Return 1Y Min"
            }
          },
          {
            "name": "max_drawdown_1y_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Drawdown 1Y Min"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Whitelisted column of funds or fund_risk_latest.",
              "default": "aum_usd",
              "title": "Sort"
            },
            "description": "Whitelisted column of funds or fund_risk_latest."
          },
          {
            "name": "dir",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "desc",
              "title": "Dir"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundsListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/funds.csv": {
      "get": {
        "tags": [
          "funds"
        ],
        "summary": "List Funds Csv",
        "description": "The same result set as /funds, unpaginated, hard-capped at 5 000 rows.",
        "operationId": "list_funds_csv_funds_csv_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 80
                },
                {
                  "type": "null"
                }
              ],
              "description": "Ticker/name substring match.",
              "title": "Search"
            },
            "description": "Ticker/name substring match."
          },
          {
            "name": "fund_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "etf",
                    "mmf",
                    "mutual_fund"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fund Type"
            }
          },
          {
            "name": "strategy_label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 80
                },
                {
                  "type": "null"
                }
              ],
              "description": "Strategy label substring match (free text).",
              "title": "Strategy Label"
            },
            "description": "Strategy label substring match (free text)."
          },
          {
            "name": "asset_class",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "equity",
                    "fixed_income",
                    "cash",
                    "alternatives"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Class"
            }
          },
          {
            "name": "expense_ratio_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expense Ratio Max"
            }
          },
          {
            "name": "aum_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aum Min"
            }
          },
          {
            "name": "sharpe_1y_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sharpe 1Y Min"
            }
          },
          {
            "name": "volatility_1y_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Volatility 1Y Max"
            }
          },
          {
            "name": "return_1y_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Return 1Y Min"
            }
          },
          {
            "name": "max_drawdown_1y_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Drawdown 1Y Min"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Whitelisted column of funds or fund_risk_latest.",
              "default": "aum_usd",
              "title": "Sort"
            },
            "description": "Whitelisted column of funds or fund_risk_latest."
          },
          {
            "name": "dir",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "desc",
              "title": "Dir"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/csv": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/funds/{instrument_id}": {
      "get": {
        "tags": [
          "funds"
        ],
        "summary": "Get Fund Profile",
        "description": "Full profile: identity + all risk metrics + 2y NAV (~260 points,\ndecimated server-side) + latest top holdings (top-50-truncated source).",
        "operationId": "get_fund_profile_funds__instrument_id__get",
        "parameters": [
          {
            "name": "instrument_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Instrument Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FundListItem": {
        "properties": {
          "instrument_id": {
            "type": "string",
            "format": "uuid",
            "title": "Instrument Id"
          },
          "series_id": {
            "type": "string",
            "title": "Series Id"
          },
          "ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticker"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "fund_type": {
            "type": "string",
            "title": "Fund Type"
          },
          "strategy_label": {
            "type": "string",
            "title": "Strategy Label"
          },
          "asset_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Class"
          },
          "is_index": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Index"
          },
          "expense_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expense Ratio"
          },
          "aum_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aum Usd"
          },
          "return_1y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return 1Y"
          },
          "volatility_1y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volatility 1Y"
          },
          "sharpe_1y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharpe 1Y"
          },
          "max_drawdown_1y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Drawdown 1Y"
          },
          "peer_sharpe_pctl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peer Sharpe Pctl"
          },
          "elite_flag": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elite Flag"
          }
        },
        "type": "object",
        "required": [
          "instrument_id",
          "series_id",
          "ticker",
          "name",
          "fund_type",
          "strategy_label",
          "asset_class",
          "is_index",
          "expense_ratio",
          "aum_usd",
          "return_1y",
          "volatility_1y",
          "sharpe_1y",
          "max_drawdown_1y",
          "peer_sharpe_pctl",
          "elite_flag"
        ],
        "title": "FundListItem",
        "description": "One row of the funds table: identity + headline risk metrics."
      },
...
      "FundProfileResponse": {
        "properties": {
          "instrument_id": {
            "type": "string",
            "format": "uuid",
            "title": "Instrument Id"
          },
          "series_id": {
            "type": "string",
            "title": "Series Id"
          },
          "ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticker"
          },
          "isin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isin"
          },
          "cusip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cusip"
          },
          "lei": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lei"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "fund_type": {
            "type": "string",
            "title": "Fund Type"
          },
          "strategy_label": {
            "type": "string",
            "title": "Strategy Label"
          },
          "asset_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Class"
          },
          "is_index": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Index"
          },
          "expense_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expense Ratio"
          },
          "aum_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aum Usd"
          },
          "primary_benchmark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Benchmark"
          },
          "inception_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inception Date"
          },
          "domicile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domicile"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "synced_at": {
            "type": "string",
            "format": "date-time",
            "title": "Synced At"
          },
          "source_calc_date": {
            "type": "string",
            "format": "date",
            "title": "Source Calc Date"
          },
          "source_nav_max_date": {
            "type": "string",
            "format": "date",
            "title": "Source Nav Max Date"
          },
          "risk": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FundRiskOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "nav": {
            "items": {
              "$ref": "#/components/schemas/FundNavPoint"
            },
            "type": "array",
            "title": "Nav"
          },
          "holdings": {
            "$ref": "#/components/schemas/FundHoldingsOut"
          },
          "classes": {
            "items": {
              "$ref": "#/components/schemas/FundClassOut"
            },
            "type": "array",
            "title": "Classes"
          },
          "classification_note": {
            "type": "string",
            "title": "Classification Note",
            "default": "Labels da fonte podem conter erros do classificador autom\u00e1tico"
          }
        },
        "type": "object",
        "required": [
          "instrument_id",
          "series_id",
          "ticker",
          "isin",
          "cusip",
          "lei",
          "name",
          "fund_type",
          "strategy_label",
          "asset_class",
          "is_index",
          "expense_ratio",
          "aum_usd",
          "primary_benchmark",
          "inception_date",
          "domicile",
          "currency",
          "synced_at",
          "source_calc_date",
          "source_nav_max_date",
          "risk",
          "nav",
          "holdings",
          "classes"
        ],
        "title": "FundProfileResponse",
        "description": "Full fund profile: identity + all risk metrics + NAV series + holdings."
      },
...
      "FundsListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FundListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "staleness": {
            "$ref": "#/components/schemas/FundsStaleness"
          },
          "classification_note": {
            "type": "string",
            "title": "Classification Note",
            "default": "Labels da fonte podem conter erros do classificador autom\u00e1tico"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "staleness"
        ],
        "title": "FundsListResponse"
      },
      "FundsStaleness": {
        "properties": {
          "synced_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synced At"
          },
          "source_calc_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Calc Date"
          },
          "source_nav_max_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Nav Max Date"
          }
        },
        "type": "object",
        "required": [
          "synced_at",
          "source_calc_date",
          "source_nav_max_date"
        ],
        "title": "FundsStaleness",
        "description": "Global data-freshness markers (max over the synced universe)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}